+++ Tickets/src/main/java/com/tickets/Tickets/entity/Comment.java
@@ -4,10 +4,10 @@ import java.io.Serializable;
 
 public class Comment implements Serializable {
     private static final long serialVersionUID = -7398187628098509372L;
-    private Integer id;
-    private Integer planid;// 所属演出id
-    private Integer userid;// 评论者id
-    private Integer parentid;// 母评论id（第一级评论默认设置为0）
+    private Long id;
+    private Long planid;// 所属演出id
+    private Long userid;// 评论者id
+    private Long parentid;// 母评论id（第一级评论默认设置为0）
     private String comment_time;// 评论时间,暂时用String表示事后需要改变
     private String content;//评论内容
     private Integer rate;// 星级
@@ -21,7 +21,7 @@ public class Comment implements Serializable {
     public Comment() {
     }
 
-    public Comment(int id, int planid, int userid, int parentid, String comment_time, String content, int rate, int type, int deleteflag) {
+    public Comment(Long id, Long planid, Long userid, Long parentid, String comment_time, String content, Integer rate, Integer type, Integer deleteflag) {
         this.id = id;
         this.planid = planid;
         this.userid = userid;
@@ -33,7 +33,7 @@ public class Comment implements Serializable {
         this.deleteflag = deleteflag;
     }
 
-    public Comment(int planid, int userid, int parentid, String comment_time, String content, int rate, int type, int deleteflag) {
+    public Comment(Long planid, Long userid, Long parentid, String comment_time, String content, Integer rate, Integer type, Integer deleteflag) {
         this.planid = planid;
         this.userid = userid;
         this.parentid = parentid;
@@ -44,35 +44,35 @@ public class Comment implements Serializable {
         this.deleteflag = deleteflag;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getPlanid() {
+    public Long getPlanid() {
         return planid;
     }
 
-    public void setPlanid(int planid) {
+    public void setPlanid(Long planid) {
         this.planid = planid;
     }
 
-    public int getUserid() {
+    public Long getUserid() {
         return userid;
     }
 
-    public void setUserid(int userid) {
+    public void setUserid(Long userid) {
         this.userid = userid;
     }
 
-    public int getParentid() {
+    public Long getParentid() {
         return parentid;
     }
 
-    public void setParentid(int parentid) {
+    public void setParentid(Long parentid) {
         this.parentid = parentid;
     }
 
@@ -92,27 +92,29 @@ public class Comment implements Serializable {
         this.content = content;
     }
 
-    public int getRate() {
+    public Integer getRate() {
         return rate;
     }
 
-    public void setRate(int rate) {
+    public void setRate(Integer rate) {
         this.rate = rate;
     }
 
-    public int getType() {
+    public Integer getType() {
         return type;
     }
 
-    public void setType(int type) {
+    public void setType(Integer type) {
         this.type = type;
     }
 
-    public int getDeleteflag() {
+    public Integer getDeleteflag() {
         return deleteflag;
     }
 
-    public void setDeleteflag(int deleteflag) {
+    public void setDeleteflag(Integer deleteflag) {
         this.deleteflag = deleteflag;
     }
 }
+
+
